Hi,
When searching for events that have not been classified, Snorby doesn't return any results. The issue was that the word null was being inserted into the search string instead of a null value. I modified the search.html.erb file to fix this.

I'm really enjoying Snorby, keep up the great work!